requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
Flask_HTTPAuth==4.5.0
Flask_SQLAlchemy==2.5.1
gunicorn==20.1.0
python_dateutil==2.8.2
SQLAlchemy==1.4.31
Werkzeug==2.0.2
